refactor: 이벤트 핸들러 코드를 관련있는 컴포넌트 파일로 이동한다
